--- Code/Source/home_page.py
@@ -1,199 +1,0 @@
-from Code.Source.globalVariables import addPage, getDataFile, getJobFile, removePage, printStack, getFirst, getLast, getUser
-import json
-
-def showHomePageGreeting():
-    printDivider()
-    print("Welcome to InCollege!")
-    print("""Please choose from one of the options below:\n1. Search for a job
-2. Find someone you know\n3. Learn a new skill\n4. Go to previously visited page\n""")
-
-def showSkillPageGreeting():
-    printDivider()
-    print("""Learn and explore the skill options below!\n1. Communication\n2. Leadership\n3. Collaboration\n4. Responsibility\n5. Time Management\n
-Please enter a number from 1-5.\nEnter x to return to the home page.\nEnter y to go to previously visited page.\n""")
-
-def showConstructionMessage(message):
-    printDivider()
-    print("We're sorry!\n'{}' feature is still under construction.\n".format(message))
-
-def homePage(): 
-    #Add home page to page stack
-    addPage("home")
-
-    showHomePageGreeting()
-    try: 
-        user_choice = int(input("Enter your option (1, 2, 3, or 4): "))
-    except:
-        print("Invalid Input!") 
-        return
-
-    while user_choice not in range(1, 5):
-        showHomePageGreeting()
-        try:
-            user_choice = int(input("Enter your option (1, 2, 3 or 4): "))   
-        except:
-            print("Invalid Input!") 
-            return
-
-    route(user_choice)
-    returnToHomePage()
-
-def route(user_choice):
-    if user_choice == 1:
-        jobPage()
-    elif user_choice == 2:
-        findSomeonePage()
-    elif user_choice == 3:
-        addPage("learnSkill")
-        showSkillPageGreeting()
-        skill_choice = input("Your choice: ")
-        while skill_choice != 'y' and skill_choice != 'x' and skill_choice not in [str(i) for i in range(1, 6)]:
-            showSkillPageGreeting()
-            skill_choice = input("Your choice: ")
-        if skill_choice == 'x':
-            homePage()
-        elif skill_choice == 'y':
-            lastPage = removePage()
-            checkPages(lastPage, homePageLinks)
-        else: 
-            skillPage(skill_choice)
-    elif user_choice == 4:
-        lastPage = removePage()
-        checkPages(lastPage, homePageLinks)
-
-#Checks all possible pages to call back to last page visited
-def checkPages(page, links):
-    if page in links:
-        links[page]()
-    else:
-        print("ERROR. Page not found")
-
-def returnToHomePage():
-    user_choice = input('Go back to homepage? (y/n): ')
-    while user_choice != 'y' and user_choice != 'n':
-        user_choice = input('Please enter y for yes, n for no: ')
-    if user_choice == 'y':
-        homePage()
-    else: 
-        printDivider()
-        exit
-
-def jobPage():
-    addPage("job")
-
-    message = "\n1. Post a job\n2. Home page\n3. Previous Page\n"
-    print(message)
-    user_choice = input("Enter your option: ")
-    while user_choice != '1' and user_choice != '2' and user_choice != '3':
-        print("\nInvalid input.\n" + message)
-        user_choice = input("Enter your option: ")
-    if user_choice == '1':
-        addJobPost()
-    elif user_choice == '2':
-        homePage() 
-    elif user_choice == '3':
-        lastPage = removePage()
-        checkPages(lastPage, homePageLinks)
-    
-def addJobPost():
-    fileName = getJobFile()
-
-    #Checks if already 5 job posts
-    with open (fileName) as jsonFile:
-        data = json.load(jsonFile)
-        temp1 = data["numPosts"]
-        if temp1 >= 5:
-            print("There are already five job posts. Try again later.")
-            return
-
-    print("Please input the following information about the job when prompted.\n")
-    while True:
-        jobTitle = input("Job Title: ")
-        length = checkLength(jobTitle, 50)
-        if length == True:
-            break
-    while True:
-        jobDescription = input("Job Description: ")
-        length = checkLength(jobDescription, 250)
-        if length == True:
-            break
-    while True:
-        jobEmployer = input("Employer: ")
-        length = checkLength(jobEmployer, 50)
-        if length == True:
-            break
-    while True:
-        jobLocation = input("Job Location: ")
-        length = checkLength(jobLocation, 50)
-        if length == True:
-            break
-    while True:
-        jobSalary = input("Job Salary: ")
-        length = checkLength(jobSalary, 50)
-        if length == True:
-            break
-    print("\n")
-
-    jobDictionary = {
-        "Title" : jobTitle,
-        "Description" : jobDescription,
-        "Employer" : jobEmployer,
-        "Location" : jobLocation,
-        "Salary" : jobSalary,
-        "Name" : getFirst() + ' ' + getLast()
-    }
-
-    #Appends new post to json file, Increase post count if < 5
-    with open (fileName) as jsonFile:
-        data = json.load(jsonFile)
-        temp = data["jobPosts"]
-        y = jobDictionary
-        temp.append(y)
-
-        #increment number of job posts
-        temp1 = data["numPosts"]
-        data["numPosts"] = temp1 + 1
-    
-    writeJson(data, fileName)
-        
-#Write data to a json File
-def writeJson(data, filename):
-    with open (filename, "w") as f:
-        json.dump(data, f, indent = 4) 
-
-#Read in job posts at application start upclear
-def readJobPosts():
-    fileName = getJobFile()
-    with open(fileName) as json_file:
-        data = json.load(json_file)
-        jobPosts = data["jobPosts"]
-    return jobPosts
-
-#Character Limiter Function (Security Measure)
-def checkLength(input, limit):
-    if len(input) > limit:
-        print("\nERROR: Maximum characters of " + str(limit) + " reached.\n")
-        return False
-    return True
-
-def findSomeonePage():
-    addPage("findSomeone")
-    showConstructionMessage("Find someone you know")
-
-def skillPage(skill):
-    showConstructionMessage("Learn a new skill")
-
-def printDivider():
-    print('\n' + '-'*60 + '\n')
-
-def main():
-    homePage()
-
-homePageLinks = {
-    "job": jobPage,
-    "findSomeone": findSomeonePage,
-    "learnSkill": skillPage,
-    "home": homePage,
-    "main": main,
-    "postJob": addJobPost
-}
