+++ testing_menu.py
@@ -4,5 +4,5 @@ from Code.Source.globalVariables import dataFileInit, stackInit
 dataFileInit()
 #Initialize stack and add login page
 stackInit()
-addPage("main")
-usefulLinksMenu()
+# addPage("main")
+incollegeImpLinks()
